requirements.txt
python-telegram-bot[webhooks]==20.6
requests
nest_asyncio